added parameters json for optimization

+++ code/portfolio_management_model.py
@@ -1,11 +1,37 @@
+import os
 import json
 from config import PORTFOLIO
 
 def implement(dec):
     if dec[0] == 0:
         return
-    if dec[0] == 1:
-        pos = 'long'
-        dictionary = {"position": pos, "amount": 50}
+    # get currant account status
+    # checks if file exists
+    if os.path.isfile(PORTFOLIO) and os.access(PORTFOLIO, os.R_OK):
+        # load account status
+        with open(PORTFOLIO, 'r') as fp:
+            account = json.load(fp)
+        if dec[0] == 1:
+            account['position'] = 'long'
+            account['amount'] = 50
+        if dec[0] == -1:
+            account['position'] = 'short'
+            account['amount'] = 50
         with open(PORTFOLIO, "w") as outfile:
-            json.dump(dictionary, outfile)
+            json.dump(account, outfile)
+    else:
+        if dec[0] == 1:
+            pos = 'long'
+            dictionary = {"position": pos, "amount": 50}
+            with open(PORTFOLIO, "w") as outfile:
+                json.dump(dictionary, outfile)
+        if dec[0] == -1:
+            pos = 'short'
+            dictionary = {"position": pos, "amount": 50}
+            with open(PORTFOLIO, "w") as outfile:
+                json.dump(dictionary, outfile)
+
+def get_results(dec,actual_price):
+    pred = dec[5]
+    diff = (actual_price - pred[50])/actual_price
+    return diff
